+++ omg-react-app/src/Home.jsx
@@ -1,12 +1,13 @@
 import React from 'react';
 import backgroundImage from './homeBackgroundImage.png'
+import logo from './logo.svg'
 import BackgroundImage from "./BackgroundImage";
-
+import Button from './Button';
+import { Route, Link, useNavigate} from 'react-router-dom';
 import styles from "./Home.module.css"
 
 import {useState} from "react";
-
-
+import { Routes } from 'react-router-dom';
 
 const OMGFontStyle = {
     fontSize : "128px",
@@ -19,28 +20,14 @@ const DescriptionFontStyle = {
     paddingLeft : "10px"
 }
 
-const ButtonFontStyle = {
-    fontSize : "24px",
-    fontWeight : "600",
-    opacity : "65%"
-}
-
-const Button = {
-    width : "324px",
-    height : "99px",
-    borderRadius : "30px",
-    boxShadow : "0px 4px 4px rgba(0,0,0,0.25)",
-    border : "none",
-    backgroundColor : "#F4EFFF"
-}
-
-const ButtonDiv = {
-    marginTop : "186px",
-    marginLeft : "20px"
-}
 // TODO
 // 버튼 클릭 시 #CEC8D8로 색깔 바꾸기
 function Home(props) {
+    const navigate = useNavigate()
+
+    const goMain = () => {
+        navigate('/main')
+    }
     return (
         <div className={styles.components}>
             <BackgroundImage url={backgroundImage}/>
@@ -48,14 +35,13 @@ function Home(props) {
                 <div>
                     <span style={OMGFontStyle}>OMG</span>
                     <p style={DescriptionFontStyle}>자신의 손글씨를 폰트로 만들어보세요.</p>
-                    <div style={ButtonDiv}>
-                        <button style={Button}>
-                            <span style={ButtonFontStyle}>폰트 만들기</span>
-                        </button>
-                    </div>
+                        <Link to={"/main"}>
+                            <Button content={"폰트 만들기"} goMain={goMain}></Button>
+                        </Link>
                 </div>
             </div>
         </div>
     );
 }
+
 export default Home;
